Tweak tests to run quicker and (hopefully) be less likely to fail on CI hardware
